QueryModel: Add attributesForURLQueryParams
Move helpers to utils.ts
