+++ packages/components/releaseNotes/labkey/components.md
@@ -12,7 +12,7 @@ Components, models, actions, and utility functions for LabKey applications and p
 * QueryModel: Added fields
     * bindURL
     * selectedReportId
-* QueryModel: Added urlQueryParams and hasRows getters
+* QueryModel: Added urlQueryParams, attributesForURLQueryParams, and hasRows getters
 * QueryModel: Default to Details View if keyValue is set
 * Removed unused getter methods from DataViewInfo
 * Added toString() to SchemaQuery class
