+++ packages/components/src/QueryModel/utils.ts
@@ -5,7 +5,7 @@
  */
 import { Filter } from '@labkey/api';
 
-import { QueryConfig, QuerySort } from '..';
+import { QuerySort } from '..';
 import { ActionValue } from '../components/omnibox/actions/Action';
 
 export function filterToString(filter: Filter.IFilter): string {
@@ -64,4 +64,35 @@ export function actionValuesToString(actionValues: ActionValue[]): string {
         .map(actionValue => actionValue.value.toString())
         .sort()
         .join(';');
+}
+
+export function offsetFromString(rowsPerPage: number, pageStr: string): number {
+    if (pageStr === undefined) {
+        return undefined;
+    }
+
+    let offset = 0;
+    const page = parseInt(pageStr, 10);
+
+    if (!isNaN(page)) {
+        offset = (page - 1) * rowsPerPage;
+    }
+
+    return offset >= 0 ? offset : 0;
+}
+
+export function querySortFromString(sortStr: string): QuerySort {
+    if (sortStr.startsWith('-')) {
+        return new QuerySort({ dir: '-', fieldKey: sortStr.slice(1) });
+    } else {
+        return new QuerySort({ fieldKey: sortStr });
+    }
+}
+
+export function querySortsFromString(sortsStr: string): QuerySort[] {
+    return sortsStr?.split(',').map(querySortFromString);
+}
+
+export function searchFiltersFromString(searchStr: string): Filter.IFilter[] {
+    return searchStr?.split(';').map(search => Filter.create('*', search, Filter.Types.Q));
 }
